Add files via upload
State Excursion Planner is a route map like system that is developed using Bellman Held Karp Algorithm using C programming. The goal of the STATE EXCURSION PLANNER is to give tourists the most efficient and optimal path possible, thus, saving the travelling time. The algorithm starts with the user selecting the journey's starting location, and then moves on to the next spot closest to the starting place.  This iterative loop continues until all of the nodes have been covered and then ends when it returns to the start point.